Move classes to separate files.
