+++ FinniversKit/Sources/Extensions/SwiftUI/View+Extensions.swift
@@ -11,4 +11,8 @@ extension View {
                 .stroke(color, lineWidth: width)
         )
     }
+
+    public func cornerRadius(_ radius: CGFloat, corners: UIRectCorner) -> some View {
+        clipShape( RoundedCorner(radius: radius, corners: corners) )
+    }
 }
