+++ FinniversKit/Sources/SwiftUI Components/RoundedCorner.swift
@@ -1,0 +1,16 @@
+import SwiftUI
+
+public struct RoundedCorner: Shape {
+    public var radius: CGFloat = .infinity
+    public var corners: UIRectCorner = .allCorners
+
+    public init(radius: CGFloat, corners: UIRectCorner) {
+        self.radius = radius
+        self.corners = corners
+    }
+
+    public func path(in rect: CGRect) -> Path {
+        let path = UIBezierPath(roundedRect: rect, byRoundingCorners: corners, cornerRadii: CGSize(width: radius, height: radius))
+        return Path(path.cgPath)
+    }
+}
