+++ FinniversKit/Sources/SwiftUI Components/SwiftUISelectionListView.swift
@@ -69,73 +69,6 @@ public struct SwiftUISelectionListView<ItemValue>: View {
     }
 }
 
-extension View {
-    public func cornerRadius(_ radius: CGFloat, corners: UIRectCorner) -> some View {
-        clipShape( RoundedCorner(radius: radius, corners: corners) )
-    }
-}
-
-public struct RoundedCorner: Shape {
-    public var radius: CGFloat = .infinity
-    public var corners: UIRectCorner = .allCorners
-
-    public init(radius: CGFloat, corners: UIRectCorner) {
-        self.radius = radius
-        self.corners = corners
-    }
-
-    public func path(in rect: CGRect) -> Path {
-        let path = UIBezierPath(roundedRect: rect, byRoundingCorners: corners, cornerRadii: CGSize(width: radius, height: radius))
-        return Path(path.cgPath)
-    }
-}
-
-public class SwiftUISelectionListItemModel<ItemValue>: ObservableObject, Identifiable {
-    public init(title: String, value: ItemValue, description: String? = nil, image: Image? = nil, isSelected: Bool) {
-        self.title = title
-        self.value = value
-        self.description = description
-        self.image = image
-        self.isSelected = isSelected
-    }
-
-    let id = UUID()
-    public let title: String
-    public let value: ItemValue
-    public let description: String?
-    public let image: Image?
-    @Published public var isSelected: Bool
-}
-
-struct SwiftUISelectionListItem<ItemType>: View {
-    @ObservedObject var itemModel: SwiftUISelectionListItemModel<ItemType>
-
-    var body: some View {
-        HStack(spacing: .spacingM) {
-            SwiftUIRadioButton(isSelected: $itemModel.isSelected)
-
-            VStack(alignment: .leading) {
-                Text(itemModel.title)
-                    .finnFont(.captionStrong)
-                    .foregroundColor(.textPrimary)
-
-                if let description = itemModel.description {
-                    Text(description)
-                        .finnFont(.caption)
-                        .foregroundColor(.textSecondary)
-                }
-            }
-            .frame(maxWidth: .infinity, alignment: .leading)
-
-            if let image = itemModel.image {
-                image
-                    .foregroundColor(.textSecondary)
-            }
-        }
-        .padding(.spacingM)
-    }
-}
-
 struct SwiftUISelectionListView_Previews: PreviewProvider {
     static var previews: some View {
         StatefulPreviewWrapper<String?, SwiftUISelectionListView<String>>("") {
@@ -166,4 +99,3 @@ struct SwiftUISelectionListView_Previews: PreviewProvider {
         }
     }
 }
-
